Replace float by int32.

+++ Source/Engine/Core/Math/Int3.h
@@ -85,7 +85,7 @@ public:
     // Init
     // @param v Int2 to use X and Y components
     // @param z Z component value
-    Int3(const Int2& xy, float z);
+    Int3(const Int2& xy, int32 z);
 
     // Init
     // @param v Int4 to use X and Y components
@@ -94,7 +94,7 @@ public:
     // Init
     // @param v Vector2 to use X and Y components
     // @param z Z component value
-    explicit Int3(const Vector2& xy, float z);
+    explicit Int3(const Vector2& xy, int32 z);
     
     // Init
     // @param v Vector3 to use X, Y and Z components
